Update:     1. Add test for audio2face     2. Fix pyproject.toml

+++ sumbody/services/audio2chunks.py
@@ -8,44 +8,40 @@ from typing import List, Tuple
 
 class Audio2Chunks:
     @staticmethod
-    def split_audio_to_chunks(audio: bytes) -> Tuple[List[bytes], int]:
+    def split_audio_to_chunks(
+        audio: bytes, audio_type: str = "mp3", framerate=None
+    ) -> Tuple[List[bytes], int]:
         """
         Split the audio into chunks
         :param audio: the audio to be split
+        :param audio_type: the audio type("mp3" or "wav"), default is "mp3".
+        :param framerate: framerate of wav file.
         :param chunk_size: the size of each chunk
         :return: a list of chunks and the sample rate
         """
-        # Generate an id for the audio files
-        audio_id = uuid.uuid4().hex
-        mp3_file = "{name}.mp3".format(name=audio_id)
-        wav_file = "{name}.wav".format(name=audio_id)
+        if audio_type == "mp3":
+            # Generate an id for the audio files
+            audio_id = uuid.uuid4().hex
+            mp3_file = "{name}.mp3".format(name=audio_id)
+            wav_file = "{name}.wav".format(name=audio_id)
 
-        # Save audio content to an mp3 file
-        with open(mp3_file, "wb") as out:
-            out.write(audio)
+            # Save audio content to an mp3 file
+            with open(mp3_file, "wb") as out:
+                out.write(audio)
 
-        # Convert MP3 to WAV
-        subprocess.call(["ffmpeg", "-i", mp3_file, wav_file])
+            # Convert MP3 to WAV
+            subprocess.call(["ffmpeg", "-i", mp3_file, wav_file])
 
-        data, sample_rate = Audio2Chunks.split_audio_to_chunks_raw(wav_file)
+            # Extract the audio data and the sample rate
+            data, sample_rate = soundfile.read(wav_file)
 
-        # Clean up the files
-        os.remove(mp3_file)
-        os.remove(wav_file)
-
-        # Return the chunks and the sample rate
-        return data, sample_rate
-    
-    @staticmethod
-    def split_audio_to_chunks_raw(audio_file: bytes) -> Tuple[List[bytes], int]:
-        """
-        Split the audio into chunks
-        :param audio: the audio file to be split (wav format)
-        :param chunk_size: the size of each chunk
-        :return: a list of chunks and the sample rate
-        """
-        # Extract the audio data and the sample rate
-        data, sample_rate = soundfile.read(audio_file)
+            # Clean up the files
+            os.remove(mp3_file)
+            os.remove(wav_file)
+        elif audio_type == "wav":
+            data, sample_rate = audio, framerate
+        else:
+            raise "Unknown datatype"
 
         # Split the audio into chunks
         chunk_size = sample_rate // 10
